Added implementation file for Pac class
