+++ CMakeLists.txt
@@ -19,5 +19,6 @@ target_sources(
   Sources/Timer.cpp
   Sources/Position.cpp
   Sources/Entity.cpp
+  Sources/Pac.cpp
 )
 target_link_libraries(pacman_sdl ${SDL2_LIBRARIES} SDL2::Main SDL2::Image SDL2::TTF SDL2::Mixer)
